+++ elasticsearch-model/lib/elasticsearch/model/adapters/mongoid.rb
@@ -9,7 +9,7 @@ module Elasticsearch
       module Mongoid
 
         Adapter.register self,
-                         lambda { |klass| !!defined?(::Mongoid::Document) && klass.ancestors.include?(::Mongoid::Document) }
+                         lambda { |klass| !!defined?(::Mongoid::Document) && klass.respond_to?(:ancestors) && klass.ancestors.include?(::Mongoid::Document) }
 
         module Records
 
@@ -66,7 +66,7 @@ module Elasticsearch
             options[:batch_size] ||= 1_000
             items = []
 
-            all.each do |item|
+            all.no_timeout.each do |item|
               items << item
 
               if items.length % options[:batch_size] == 0