+++ elasticsearch-rails/test/unit/instrumentation/lograge_test.rb
@@ -1,6 +1,7 @@
 require 'test_helper'
 
 require 'rails/railtie'
+require 'action_pack'
 require 'lograge'
 
 require 'elasticsearch/rails/lograge'